Collected small changes for the next patch release

+++ src/KSPACE/msm_cg.cpp
@@ -545,6 +545,6 @@ void MSMCG::fieldforce_peratom()
 double MSMCG::memory_usage()
 {
   double bytes = MSM::memory_usage();
-  bytes += nmax * sizeof(int);
+  bytes += (double)nmax * sizeof(int);
   return bytes;
 }
