+++ src/MANYBODY/pair_eim.cpp
@@ -1045,8 +1045,8 @@ void PairEIM::unpack_reverse_comm(int n, int *list, double *buf)
 double PairEIM::memory_usage()
 {
   double bytes = maxeatom * sizeof(double);
-  bytes += maxvatom*6 * sizeof(double);
-  bytes += 2 * nmax * sizeof(double);
+  bytes += (double)maxvatom*6 * sizeof(double);
+  bytes += (double)2 * nmax * sizeof(double);
   return bytes;
 }
