+++ src/MC/fix_bond_create.cpp
@@ -1429,7 +1429,7 @@ double FixBondCreate::memory_usage()
   int nmax = atom->nmax;
   double bytes = nmax * sizeof(int);
   bytes = 2*nmax * sizeof(tagint);
-  bytes += nmax * sizeof(double);
+  bytes += (double)nmax * sizeof(double);
   return bytes;
 }
