+++ src/POEMS/fix_poems.cpp
@@ -1526,8 +1526,8 @@ double FixPOEMS::memory_usage()
 {
   int nmax = atom->nmax;
   double bytes = nmax * sizeof(int);
-  bytes += nmax*MAXBODY * sizeof(int);
-  bytes += nmax*3 * sizeof(double);
+  bytes += (double)nmax*MAXBODY * sizeof(int);
+  bytes += (double)nmax*3 * sizeof(double);
   return bytes;
 }
