+++ src/USER-LB/fix_lb_rigid_pc_sphere.cpp
@@ -1431,8 +1431,8 @@ double FixLbRigidPCSphere::memory_usage()
 {
   int nmax = atom->nmax;
   double bytes = nmax * sizeof(int);
-  bytes += nmax*3 * sizeof(double);
-  bytes += maxvatom*6 * sizeof(double);
+  bytes += (double)nmax*3 * sizeof(double);
+  bytes += (double)maxvatom*6 * sizeof(double);
 
   return bytes;
 }
