+++ src/USER-OMP/pair_brownian_poly_omp.cpp
@@ -396,8 +396,8 @@ double PairBrownianPolyOMP::memory_usage()
 {
   double bytes = memory_usage_thr();
   bytes += PairBrownianPoly::memory_usage();
-  bytes += nthreads * sizeof(RanMars*);
-  bytes += nthreads * sizeof(RanMars);
+  bytes += (double)nthreads * sizeof(RanMars*);
+  bytes += (double)nthreads * sizeof(RanMars);
 
   return bytes;
 }
