+++ src/USER-SDPD/fix_meso_move.cpp
@@ -800,8 +800,8 @@ void FixMesoMove::final_integrate () {
 
 double FixMesoMove::memory_usage () {
   double bytes = atom->nmax*3 * sizeof(double);
-  if (displaceflag) bytes += atom->nmax*3 * sizeof(double);
-  if (velocityflag) bytes += atom->nmax*3 * sizeof(double);
+  if (displaceflag) bytes += (double)atom->nmax*3 * sizeof(double);
+  if (velocityflag) bytes += (double)atom->nmax*3 * sizeof(double);
   return bytes;
 }
