+++ src/compute_gyration_chunk.cpp
@@ -357,8 +357,8 @@ void ComputeGyrationChunk::allocate()
 double ComputeGyrationChunk::memory_usage()
 {
   double bytes = (bigint) maxchunk * 2 * sizeof(double);
-  bytes += (bigint) maxchunk * 2*3 * sizeof(double);
-  if (tensor) bytes += (bigint) maxchunk * 2*6 * sizeof(double);
-  else bytes += (bigint) maxchunk * 2 * sizeof(double);
+  bytes += (double) maxchunk * 2*3 * sizeof(double);
+  if (tensor) bytes += (double) maxchunk * 2*6 * sizeof(double);
+  else bytes += (double) maxchunk * 2 * sizeof(double);
   return bytes;
 }
