+++ src/compute_orientorder_atom.cpp
@@ -332,8 +332,8 @@ void ComputeOrientOrderAtom::compute_peratom()
 double ComputeOrientOrderAtom::memory_usage()
 {
   double bytes = ncol*nmax * sizeof(double);
-  bytes += (qmax*(2*qmax+1)+maxneigh*4) * sizeof(double);
-  bytes += (nqlist+maxneigh) * sizeof(int);
+  bytes += (double)(qmax*(2*qmax+1)+maxneigh*4) * sizeof(double);
+  bytes += (double)(nqlist+maxneigh) * sizeof(int);
   return bytes;
 }
