+++ src/pair_coul_streitz.cpp
@@ -737,9 +737,9 @@ void PairCoulStreitz::ewald_sum(double qi, double qj, double zj, double r,
 double PairCoulStreitz::memory_usage()
 {
   double bytes = maxeatom * sizeof(double);
-  bytes += maxvatom*6 * sizeof(double);
-  bytes += nmax * sizeof(int);
-  bytes += MAXNEIGH * nmax * sizeof(int);
+  bytes += (double)maxvatom*6 * sizeof(double);
+  bytes += (double)nmax * sizeof(int);
+  bytes += (double)MAXNEIGH * nmax * sizeof(int);
   return bytes;
 }
